+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog] and this project adheres to [Semantic Versioning].
 
 
+## unreleased
+
+### Added
+- `no-property-access` rule
+
+
 ## [1.1.1] (2017-06-01)
 
 ### Added